Fix gcc build errors

+++ core/coreobjects/src/permission_manager_impl.cpp
@@ -46,7 +46,7 @@ ErrCode INTERFACE_FUNC PermissionManagerImpl::isAuthorized(IUser* user, Permissi
 
     for (const auto& group : groups)
     {
-        permissionMask = permissions.getDenied().hasKey(group) ? permissions.getDenied().get(group) : 0;
+        permissionMask = permissions.getDenied().hasKey(group) ? (Int) permissions.getDenied().get(group) : 0;
 
         if ((permissionMask & targetPermissionInt) != 0)
         {
@@ -57,7 +57,7 @@ ErrCode INTERFACE_FUNC PermissionManagerImpl::isAuthorized(IUser* user, Permissi
 
     for (const auto& group : groups)
     {
-        permissionMask = permissions.getAllowed().hasKey(group) ? permissions.getAllowed().get(group) : 0;
+        permissionMask = permissions.getAllowed().hasKey(group) ? (Int) permissions.getAllowed().get(group) : 0;
 
         if ((permissionMask & targetPermissionInt) != 0)
         {
